feat: Refactor social links to use array structure
Changes the site configuration and rendering of social links from a simple object with predefined keys (facebook, twitter, linkedin) to a more flexible array of `SocialLink` objects. This allows for arbitrary social platforms to be added and managed, improving the extensibility of the site configuration.

The `ContactPage` and `Footer` components have been updated to consume this new array structure. The `AdminDashboard` now includes functionality to add, edit, and remove social links. Default content has also been updated to reflect the new structure.